refactor(codegen): decompose events check by concern
