+++ scripts/codegen/src/plugins/events/check/index.ts
@@ -1,0 +1,2 @@
+export { checkEvents } from "./check-events.ts";
+export { checkEventsRuntimeSupport } from "./runtime-support.ts";
